+++ src/main/java/roomescape/controller/dto/request/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.request;
 
 import java.time.LocalDate;
