+++ src/main/java/roomescape/controller/dto/request/ReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.request;
 
 import java.time.LocalTime;
 import roomescape.domain.ReservationTime;
