+++ src/main/java/roomescape/controller/dto/request/ThemeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.request;
 
 public record ThemeRequest(String name, String description, String thumbnail) {
 }
