+++ src/main/java/roomescape/controller/dto/response/ReservationResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.controller.dto.response;
 
 import java.time.LocalDate;
 import roomescape.domain.Reservation;
