+++ src/test/java/roomescape/e2e/ReservationTimeApiE2ETest.java
@@ -1,5 +1,7 @@
 package roomescape.e2e;
 
+import static org.hamcrest.CoreMatchers.is;
+
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import java.time.LocalTime;
@@ -7,9 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.controller.dto.ReservationTimeRequest;
-
-import static org.hamcrest.CoreMatchers.is;
+import roomescape.controller.dto.request.ReservationTimeRequest;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
